pyproject.toml
[tool.poetry]
name = "livekit-token-generate-api"
version = "0.1.0"
description = ""
authors = ["narendra-bluebash <narendra.dangi@bluebash.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
livekit-api = "^0.8.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
